ochestration/requirements.txt
apache-airflow
apache-airflow-providers-google
pandas
google-cloud-storage
kagglehub[pandas-datasets]
